# contracts/http/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis_plan_version": {
            "title": "Analysis Plan Version",
            "type": "string"
          },
          "analyzer_results": {
            "items": {
              "$ref": "#/components/schemas/AnalyzerResultResponse"
            },
            "title": "Analyzer Results",
            "type": "array"
          },
          "assessed_at": {
            "format": "date-time",
            "title": "Assessed At",
            "type": "string"
          },
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "completeness": {
            "title": "Completeness",
            "type": "string"
          },
          "context_version": {
            "title": "Context Version",
            "type": "integer"
          },
          "has_mock_evidence": {
            "title": "Has Mock Evidence",
            "type": "boolean"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "score_kind": {
            "title": "Score Kind",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalResponse"
            },
            "title": "Signals",
            "type": "array"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "assessment_id",
          "target_id",
          "severity",
          "score",
          "score_kind",
          "completeness",
          "reason_codes",
          "signals",
          "analyzer_results",
          "policy_version",
          "analysis_plan_version",
          "assessed_at",
          "context_version",
          "has_mock_evidence"
        ],
        "title": "AssessmentResponse",
        "type": "object"
      },
...
      "ContactAnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis_pending": {
            "title": "Analysis Pending",
            "type": "boolean"
          },
          "assessment": {
            "$ref": "#/components/schemas/AssessmentResponse"
          },
          "duplicate_source_event": {
            "title": "Duplicate Source Event",
            "type": "boolean"
          },
          "incident_context_version": {
            "title": "Incident Context Version",
            "type": "integer"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "warning_dispatched": {
            "title": "Warning Dispatched",
            "type": "boolean"
          },
          "warning_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning Id"
          }
        },
        "required": [
          "observation_id",
          "incident_id",
          "incident_context_version",
          "analysis_pending",
          "assessment",
          "warning_id",
          "warning_dispatched",
          "duplicate_source_event"
        ],
        "title": "ContactAnalysisResponse",
        "type": "object"
      },
...
      "GuidanceResponse": {
        "additionalProperties": false,
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/AllowedActionResponse"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "completeness": {
            "title": "Completeness",
            "type": "string"
          },
          "content_version": {
            "title": "Content Version",
            "type": "string"
          },
          "education_cards": {
            "items": {
              "$ref": "#/components/schemas/EducationCardReferenceResponse"
            },
            "title": "Education Cards",
            "type": "array"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "requested_locale": {
            "title": "Requested Locale",
            "type": "string"
          },
          "risk_label": {
            "title": "Risk Label",
            "type": "string"
          },
          "support_contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrustedSupportContactResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "support_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Message"
          },
          "unmapped_reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Unmapped Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "risk_label",
          "explanation",
          "classification",
          "completeness",
          "recommendations",
          "education_cards",
          "allowed_actions",
          "support_message",
          "support_contact",
          "unmapped_reason_codes",
          "locale",
          "requested_locale",
          "content_version"
        ],
        "title": "GuidanceResponse",
        "type": "object"
      },
      "IncidentResponse": {
        "additionalProperties": false,
        "properties": {
          "assessment_ids": {
            "items": {
              "type": "string"
            },
            "title": "Assessment Ids",
            "type": "array"
          },
          "context_version": {
            "title": "Context Version",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "execution_mode": {
            "title": "Execution Mode",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "latest_assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Assessment Id"
          },
          "observation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Observation Ids",
            "type": "array"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItemResponse"
            },
            "title": "Timeline",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "status",
          "observation_ids",
          "assessment_ids",
          "latest_assessment_id",
          "resolution",
          "context_version",
          "revision",
          "created_at",
          "updated_at",
          "execution_mode",
          "timeline"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
...
      "ManualObservationInputSchema": {
        "discriminator": {
          "mapping": {
            "call_transcript": "#/components/schemas/CallTranscriptObservationInputSchema",
            "messenger": "#/components/schemas/MessengerObservationInputSchema",
            "sms": "#/components/schemas/SmsObservationInputSchema",
            "web_resource": "#/components/schemas/ManualWebResourceObservationInputSchema"
          },
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/SmsObservationInputSchema"
          },
          {
            "$ref": "#/components/schemas/MessengerObservationInputSchema"
          },
          {
            "$ref": "#/components/schemas/CallTranscriptObservationInputSchema"
          },
          {
            "$ref": "#/components/schemas/ManualWebResourceObservationInputSchema"
          }
        ],
        "title": "ManualObservationInputSchema"
      },
...
      "ObservationResponse": {
        "additionalProperties": false,
        "properties": {
          "execution_mode": {
            "title": "Execution Mode",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "normalization_version": {
            "title": "Normalization Version",
            "type": "string"
          },
          "normalized_indicators": {
            "items": {
              "$ref": "#/components/schemas/NormalizedIndicatorResponse"
            },
            "title": "Normalized Indicators",
            "type": "array"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SmsPayloadSchema"
              },
              {
                "$ref": "#/components/schemas/CallTranscriptPayloadSchema"
              },
              {
                "$ref": "#/components/schemas/WebResourcePayloadSchema"
              }
            ],
            "title": "Payload"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "string"
          }
        },
        "required": [
          "observation_id",
          "kind",
          "source_event_id",
          "occurred_at",
          "received_at",
          "payload",
          "normalized_indicators",
          "normalization_version",
          "execution_mode"
        ],
        "title": "ObservationResponse",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "RFC 9457 response extended with stable application fields.",
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "instance": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Synthetic-data anti-fraud prototype API.",
    "title": "Alpha Defense API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assessments/{assessment_id}": {
      "get": {
        "operationId": "get_assessment",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Get Assessment",
        "tags": [
          "contact"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/guidance": {
      "get": {
        "operationId": "get_assessment_guidance",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "ru-RU",
              "pattern": "^[a-z]{2}-[A-Z]{2}$",
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Get Assessment Guidance",
        "tags": [
          "contact"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "contact"
        ]
      }
    },
    "/api/v1/observations": {
      "post": {
        "operationId": "submit_observation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualObservationInputSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Submit Observation",
        "tags": [
          "contact"
        ]
      }
    },
    "/api/v1/observations/{observation_id}": {
      "get": {
        "operationId": "get_observation",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Get Observation",
        "tags": [
          "contact"
        ]
      }
    },
    "/api/v1/observations/{observation_id}/reassess": {
      "post": {
        "operationId": "reassess_observation",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details"
          }
        },
        "summary": "Reassess Observation",
        "tags": [
          "contact"
        ]
      }
    },
...
  }
}